Trying out smooth movement with gamepads. Fix item buying race condition

+++ src/gamestate/GameAction.cpp
@@ -16,7 +16,7 @@ GameAction::GameAction() :
 }
 
 GameAction::GameAction(obj_id origin, action_t act, int frame, int level,
-		int use_id, int action_x, int action_y, int use_id2) :
+		int use_id, float action_x, float action_y, int use_id2) :
 		origin(origin), act(act), frame(frame), room(level), use_id(use_id), use_id2(
 				use_id2), action_x(action_x), action_y(action_y) {
 
@@ -25,7 +25,7 @@ void to_action_file(FILE* f, const GameAction& action) {
 	fwrite(&action, sizeof(GameAction), 1, f);
 }
 GameAction game_action(GameState* gs, GameInst* origin,
-		GameAction::action_t action, int use_id, int action_x, int action_y,
+		GameAction::action_t action, int use_id, float action_x, float action_y,
 		int use_id2) {
 	return GameAction(origin->id, action, gs->frame(),
 			gs->get_level()->id(), use_id, action_x, action_y, use_id2);
